% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{compileResults}
\alias{compileResults}
\title{Compile results of calculation.}
\usage{
compileResults(file)
}
\arguments{
\item{file}{The data frame being analysed containing cash flow time series for
any number of investment projects.}
}
\value{
A data frame containing the IRR values for each investment.
}
\description{
Compile continuous, yearly and monthly IRR in data frame.
}
