% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{convertDatesToMonthAndY}
\alias{convertDatesToMonthAndY}
\title{Order dates as number of whole months since start.}
\usage{
convertDatesToMonthAndY(dates)
}
\arguments{
\item{dates}{A list of dates.}
}
\value{
The order of dates in months since start in integer format.
}
\description{
Values express to which month in order the date in the dataset belongs.
}
\examples{
convertDatesToMonthAndY(c("1.1.2000", "1.3.2002")) would return 1 and 27.
}
