% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{convertDatesToT}
\alias{convertDatesToT}
\title{Convert dates in a list to a floating point value.}
\usage{
convertDatesToT(dates)
}
\arguments{
\item{dates}{A list of dates.}
}
\value{
The dates in years and fractions of years in floating point format.
}
\description{
The values express time elapsed since
the first date in the dataset used and is measured in whole years, with
days and months expressed as fractions of a year.
}
\details{
Note that the day count convention used is 30E/360.
}
\examples{
convertDatesToT(c("1.1.2000", "1.7.2002")) will return 0.0 and 2.5.
}
