% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{convertDatesToY}
\alias{convertDatesToY}
\title{Order dates as number of whole years since start.}
\usage{
convertDatesToY(dates)
}
\arguments{
\item{dates}{A list of dates.}
}
\value{
The order of dates in years since start in integer format.
}
\description{
Values express to which year in order the date in the dataset belongs.
}
\examples{
convertDatesToMonthAndY(c("1.1.2000", "1.3.2002")) would return 1 and 3.
}
