% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{convertVectorToFunction}
\alias{convertVectorToFunction}
\title{Convert a vector of cashflows into a continuous discount function}
\usage{
convertVectorToFunction(irr, vect, dates)
}
\arguments{
\item{irr}{The unknown IRR variable to be solved for.}

\item{vect}{A vector with the cashflows to be analysed.}

\item{dates}{A vector with dates as fractions of years corresponding to the cashflows.}
}
\value{
A function for the NPV in the current case with an unknown
interest rate to be solved for.
}
\description{
The continuous discount function calculates NPV as the sum of the following
for each cash flow CF at point in time t:
e^(-rt)
where r is the interest rate. The IRR is the interest rate for which the NPV = 0.
}
