% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{findFirst}
\alias{findFirst}
\title{Find the first index with a non-zero value in a vector.}
\usage{
findFirst(vect)
}
\arguments{
\item{vect}{A vector.}
}
\value{
The first index with a non-zero value in the vector.
}
\description{
Find the first index with a non-zero value in a vector.
}
\examples{
findFirst(c(0,0,1,0,4,2))
findFirst(c(0,6,3,4))
}
