% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{findRootContinuous}
\alias{findRootContinuous}
\title{Solve for the IRR in the continuous case}
\usage{
findRootContinuous(vect, dates)
}
\arguments{
\item{vect}{A vector with the cashflows to be analysed.}

\item{dates}{A vector with dates as fractions of years.}
}
\value{
The IRR for cashflows provided.
}
\description{
Solve for the IRR in the continuous case
}
