% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{findRootMonthly}
\alias{findRootMonthly}
\title{Solve for the IRR with monthly aggregation of cash flows.}
\usage{
findRootMonthly(vect, dates)
}
\arguments{
\item{vect}{A vector with the cashflows to be analysed.}

\item{dates}{A vector with dates months from start.}
}
\value{
The IRR for the cashflows provided.
}
\description{
Solve for the IRR with monthly aggregation of cash flows.
}
