% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{findRootYearly}
\alias{findRootYearly}
\title{Solve for the IRR with yearly aggregation of cash flows.}
\usage{
findRootYearly(vect, dates)
}
\arguments{
\item{vect}{A vector with the cashflows to be analysed.}

\item{dates}{A vector with dates as years from start.}
}
\value{
The IRR for the cashflows provided.
}
\description{
Solve for the IRR with yearly aggregation of cash flows.
}
