% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_calculator.R
\name{readFile}
\alias{readFile}
\title{Read .csv file into data frame.}
\usage{
readFile(fileName)
}
\arguments{
\item{fileName}{the name of the file to be read if in the working directory, else
complete URL of the file.}
}
\value{
Data frame adapted to the needs of the program.
}
\description{
Reads the data to be analysed from a .csv file into a data frame, then
reshapes it into a format more suitable for the calculations.
}
